fixed all issues
