+++ routes/v1/index.ts
@@ -2,9 +2,10 @@ import express from "express"
 import homeRouter from ".."
 import authRouter from "../auth"
 import payRouter from "../payment"
-import settingsRouter from "../settings"
 import dashboardRouter from "../dashboard"
 import transactionRouter from "../transactions"
+import userRouter from "../user"
+import walletRouter from "../wallets"
 
 const allRoutes = express()
 
@@ -12,7 +13,8 @@ allRoutes.use('/home', homeRouter)
 allRoutes.use('/auth', authRouter)
 allRoutes.use('/dashboard', dashboardRouter)
 allRoutes.use('/payments', payRouter)
-allRoutes.use('/settings', settingsRouter)
+allRoutes.use('/wallets', walletRouter)
 allRoutes.use('/transactions', transactionRouter)
+allRoutes.use('/user', userRouter)
 
 export default allRoutes
